feat: enforce cart stock limits, support decimal quantities, and restore stock on transaction delete

+++ backend/main.py
@@ -53,8 +53,9 @@ class CheckoutItem(BaseModel):
     id: str
     name: str
     price: float
-    qty: int
+    qty: float
     unit: str
+
 
 class CheckoutRequest(BaseModel):
     tx_id: str
